+++ apps/inference/models/one_mean_statistic.cpp
@@ -6,7 +6,7 @@
 namespace Inference {
 
 void OneMeanStatistic::computeParametersFromSeries(
-    const InferenceModel* inference, int pageIndex) {
+    const InferenceModel* inference, uint8_t pageIndex) {
   assert(hasSeries(pageIndex));
   int seriesIndex = seriesAt(pageIndex);
   Shared::StoreToSeries seriesModel(this, seriesIndex);
