+++ apps/inference/models/slope_t_statistic.cpp
@@ -8,7 +8,7 @@
 namespace Inference {
 
 void SlopeTStatistic::computeParametersFromSeries(
-    const InferenceModel* inference, int pageIndex) {
+    const InferenceModel* inference, uint8_t pageIndex) {
   assert(hasSeries(pageIndex));
   int seriesIndex = seriesAt(pageIndex);
   Shared::StoreToSeries seriesModel(this, seriesIndex);
